requirements.txt
aiogram==3.22.0
psycopg2-binary
python-dotenv
